requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0

# Music features (optional - uncomment if needed)
# yt-dlp>=2023.12.30
# PyNaCl>=1.5.0
# spotipy>=2.23.0  # For Spotify API support
# psutil>=5.9.0  # For resource monitoring

